requirements.txt
fastapi~=0.115.5
requests~=2.32.3
python-dotenv~=1.0.1
google~=3.0.0